refactor: use same map component for both settings page and school profile page
